FEAT: Cada Coleta possui uma pontuação

+++ src/main/java/com/gs/ecoocean/dto/coleta/ColetaCreateDTO.java
@@ -1,11 +1,13 @@
 package com.gs.ecoocean.dto.coleta;
 
+import java.math.BigDecimal;
+
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 
 public record ColetaCreateDTO(
         @NotNull Integer tipoLixo,
-        @NotNull @Positive Integer quantidade,
+        @NotNull @Positive BigDecimal quantidade,
         @NotNull Long idParticipacao
 ) {
 }
